+++ Cargo.toml
@@ -211,7 +211,8 @@ members = [
   "src/parquet2",
   "src/parquet-format-safe",
   "src/generated/spark-connect",
-  "src/daft-cli"
+  "src/daft-cli",
+  "src/common/progress-bar"
 ]
 
 [workspace.dependencies]
